backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
bcrypt==4.1.2
python-multipart==0.0.6
httpx==0.25.2